backend/requirements.txt
# Quizly Backend Dependencies
# FastAPI + PostgreSQL + AI

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
aiosqlite==0.22.1
alembic==1.14.0
greenlet==3.1.1

# Rate Limiting
slowapi==0.1.9

# Validation
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0

# AI - Gemini
google-generativeai==0.8.0

# HTTP
httpx==0.27.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utils
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
